+++ face-analyzer/src/__tests__/charts.test.js
@@ -1,28 +1,29 @@
 import {useAuth} from '../context/authContext';
 import {MemoryRouter, useNavigate} from 'react-router-dom';
-import Stats from '../views/stimuli/Stimuli';
-import {act, render, waitFor} from '@testing-library/react';
+import {render, waitFor} from '@testing-library/react';
 import React from 'react';
 import axios from 'axios';
+import Stats from "../views/stimuli/charts/Stats";
 
 axios.get = jest.fn();
-
-const mockStream = {
-  getTracks: jest.fn(() => [{stop: jest.fn()}]),
-};
-
-const mediaDevicesMock = {
-  getUserMedia: jest.fn(() => {
-    return Promise.resolve(mockStream)
-  }),
-};
-
-global.navigator.mediaDevices = mediaDevicesMock;
 
 const mockedUseNavigate = jest.fn();
 const mockedUseParams = jest.fn(() => {
   reactionId: '151';
 });
+
+class MockResizeObserver {
+  observe() {
+  }
+
+  unobserve() {
+  }
+
+  disconnect() {
+  }
+}
+
+window.ResizeObserver = MockResizeObserver;
 
 jest.mock('../context/authContext');
 jest.mock('react-router-dom', () => ({
@@ -34,7 +35,7 @@ jest.mock('react-router', () => ({
   useParams: () => mockedUseParams
 }));
 
-describe.skip('Charts Component', () => {
+describe('Charts Component', () => {
   describe('User is not authenticated', () => {
     beforeEach(() => {
       useAuth.mockReturnValue({ user: null });
@@ -65,86 +66,20 @@ describe.skip('Charts Component', () => {
 
     test('Chart header should be rendered for Admin', async () => {
       axios.get.mockResolvedValue(response);
-      let getByText;
 
-      await act(async () => {
-        getByText = render(
-            <MemoryRouter>
-              <Stats/>
-            </MemoryRouter>
-        ).getByText(/emotions over time/i);
-      });
-
-      await waitFor(() => expect(getByText.toBeInTheDocument()));
-      // await waitFor(() => expect(getByText(/emotions distribution/i)).toBeInTheDocument());
-      // await waitFor(() => expect(getByText(/dynamic chart/i)).toBeInTheDocument());
-      // await waitFor(() => expect(getByText(/add note/i)).toBeInTheDocument());
-      // await waitFor(() => expect(getByText(/export csv/i)).toBeInTheDocument());
-    });
-
-    test('Emotions over time should be rendered for Admin', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
+      const {getByText} = render(
           <MemoryRouter>
-            <Stats />
+            <Stats/>
           </MemoryRouter>
-        );
+      );
 
-        await waitFor(() => expect(getByText(/emotion level/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/anger/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/disgust/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/fear/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/happiness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/sadness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/surprise/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/neutral/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/video time in ms/i)).toBeInTheDocument());
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
+      await waitFor(() => expect(getByText(/emotions over time/i)).toBeInTheDocument());
+      await waitFor(() => expect(getByText(/emotions distribution/i)).toBeInTheDocument());
+      await waitFor(() => expect(getByText(/dynamic chart/i)).toBeInTheDocument());
+      await waitFor(() => expect(getByText(/add note/i)).toBeInTheDocument());
+      await waitFor(() => expect(getByText(/export csv/i)).toBeInTheDocument());
     });
 
-    test('Typography should be rendered for Admin, if activeButton is "distribution"', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
-          <MemoryRouter>
-            <Stats />
-          </MemoryRouter>
-        );
-
-        await waitFor(() => expect(getByText(/reaction duration/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/emotions distribution/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/legend/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/For more information on box plot charts check out one of the following links/i)).toBeInTheDocument());
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
-    });
-
-    test('DynamicChart should be rendered for Admin, if activeButton is "dynamic"', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
-          <MemoryRouter>
-            <Stats />
-          </MemoryRouter>
-        );
-
-        await waitFor(() => expect(getByText(/emotion level/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/anger/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/disgust/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/fear/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/happiness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/sadness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/surprise/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/neutral/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/video time in ms/i)).toBeInTheDocument());
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
-    });
   });
 
   describe('User is authenticated as Researcher', () => {
@@ -156,7 +91,6 @@ describe.skip('Charts Component', () => {
 
     test('Chart header should be rendered for Researcher', async () => {
       axios.get.mockResolvedValue(response);
-      try {
         const { getByText } = render(
           <MemoryRouter>
             <Stats />
@@ -169,76 +103,7 @@ describe.skip('Charts Component', () => {
         await waitFor(() => expect(getByText(/add note/i)).toBeInTheDocument());
         await waitFor(() => expect(getByText(/export csv/i)).toBeInTheDocument());
 
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
     });
 
-    test('Emotions over time should be rendered for Researcher', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
-          <MemoryRouter>
-            <Stats />
-          </MemoryRouter>
-        );
-
-        await waitFor(() => expect(getByText(/emotion level/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/anger/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/disgust/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/fear/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/happiness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/sadness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/surprise/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/neutral/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/video time in ms/i)).toBeInTheDocument());
-
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
-    });
-
-    test('Typography should be rendered for Researcher, if activeButton is "distribution"', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
-          <MemoryRouter>
-            <Stats />
-          </MemoryRouter>
-        );
-
-        await waitFor(() => expect(getByText(/reaction duration/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/emotions distribution/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/legend/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/For more information on box plot charts check out one of the following links/i)).toBeInTheDocument());
-
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
-    });
-
-    test('DynamicChart should be rendered for Researcher, if activeButton is "dynamic"', async () => {
-      axios.get.mockResolvedValue(response);
-      try {
-        const { getByText } = render(
-          <MemoryRouter>
-            <Stats />
-          </MemoryRouter>
-        );
-
-        await waitFor(() => expect(getByText(/emotion level/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/anger/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/disgust/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/fear/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/happiness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/sadness/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/surprise/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/neutral/i)).toBeInTheDocument());
-        await waitFor(() => expect(getByText(/video time in ms/i)).toBeInTheDocument());
-
-      } catch (e) {
-        expect(e.message).toContain('ResizeObserver is not defined');
-      }
-    });
   });
 });
